requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
altair>=5.3.0
requests>=2.31.0
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.0.0
google-api-python-client>=2.100.0
openai>=1.0.0
